requirements.txt
boto3==1.11.11
Django==2.2.10
djangorestframework==3.11.0
django-extensions==2.2.6
django-redis==4.11.0
django-storages==1.9.1
django-widget-tweaks==1.4.5
dj-database-url==0.5.0
gunicorn==19.9.0
factory-boy==2.12.0
psycopg2==2.8.4
redis==3.4.1
whitenoise==4.1.2